% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookdownbr_skeleton.R
\name{bookdownbr_skeleton}
\alias{bookdownbr_skeleton}
\title{bookdownbr_skeleton}
\usage{
bookdownbr_skeleton(path)
}
\arguments{
\item{path}{path}
}
\description{
bookdownbr_skeleton
}
\examples{
bookdownbr_skeleton(path = "c:/bookbr/brazil")
}
