% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeleton_bookdownbr.R
\name{skeleton_bookdownbr}
\alias{skeleton_bookdownbr}
\title{skeleton_bookdownbr}
\usage{
skeleton_bookdownbr(path, ...)
}
\arguments{
\item{path}{local em que sera criado o bookdown}

\item{...}{more}
}
\description{
Arquivos basicos para bookdown em portugues
}
\examples{
bookdownbr::skeleton_bookdownbr(path = "c:/bookbr/brazil")
}
